Add support for finding references using a remote fs
For find refs, we needed to patch the get_modules_containing_name function, which accessed the local file system directly to search for files containing references to a python name. Two optional arguments were added to the evaluator object: a function to list pathnames of all modules in a workspace and a function to read the contents of a path. A more architected solution may be to provide an optional bfs object which could generically satisfy these two operations across all functions, but for first steps this simple approach works well.